blogs/requirements.txt
asgiref==3.5.2
boto3==1.24.21
botocore==1.27.21
Django==3.2
django-environ==0.9.0
django-filter==22.1
django-fsm==2.8.0
django-model-utils==4.2.0
django-phonenumber-field==6.3.0
django-render-block==0.9.1
django-storages==1.12.3
django-templated-email==3.0.0
djangorestframework==3.13.1
djangorestframework-recursive==0.1.2
drf-nested-routers==0.93.4
drf-secure-token==1.1.0
jmespath==1.0.1
mysqlclient==2.1.1
phonenumbers==8.12.50
Pillow==9.1.1
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
unicef-djangolib==0.6
unicef-restlib==0.7
urllib3==1.26.9
